+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support `delete` command when accessing a remote cluster deployed on kubernetes
 
+### Fixed
+- Filter -C and --config-file option in kubernetes context
+
 ## 1.19.0 - 2021-07-26
 
 ### Added
@@ -16,7 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Calls to the AVRO schema registry are now cached correctly
-
 
 ## 1.18.1 - 2021-07-15